requirements.txt
langchain
pymongo
bs4
tiktoken
gradio
lxml
requests
argparse
unstructured
sentence-transformers
python-dotenv